[paypal-refactor] Removed functions from PayPal actions and set them up in PayPal module.
